feat: Provide ability to create statefulset
When using verdaccio as purely an edge cache, multiple deployment replicas require that
the storage PV support ReadWriteMany. However, such PVs are not always available.

A StatefulSet can be used instead, where each replica will have it's own PV.
This has the downside that initially the storage will be empty and may take time
for them to get populated.

To be clear, Using StatefulSet is primarly useful for when verdaccio is an edge
cache; Using it as a regular NPM registry (accepting uploads), would require
additional step of syncing artifacts.